Move legacy cache and caskroom code to `compat/*`.
