+++ Library/Homebrew/cask/lib/hbc/cache.rb
@@ -8,13 +8,5 @@ module Hbc
       odebug "Creating Cache at #{Hbc.cache}"
       Hbc.cache.mkpath
     end
-
-    def delete_legacy_cache
-      legacy_cache = HOMEBREW_CACHE.join("Casks")
-      return unless legacy_cache.exist?
-
-      ohai "Deleting legacy cache at #{legacy_cache}..."
-      FileUtils.remove_entry_secure(legacy_cache)
-    end
   end
 end
